chore: iOS and JS interfaces
