+++ ios/Configuration/ApplePay/ApplepayConfigurationParser.swift
@@ -104,6 +104,49 @@ public struct ApplepayConfigurationParser {
         return shippingMethods.isEmpty ? nil : shippingMethods
     }
 
+        // A description of the recurring payment, for example "Apple News+".
+    var paymentDescription: String? {
+        return dict[ApplePayKeys.Recurring.paymentDescription] as? String
+    }
+
+        // The regular billing cycle, for example "$9.99 monthly".
+    @available(iOS 15.0, *)
+    var regularBilling: PKRecurringPaymentSummaryItem? {
+        guard let dictionary = dict[ApplePayKeys.Recurring.regularBilling] as? [String: Any] else {
+            return nil
+        }
+        return .init(dictionary)
+    }
+
+        // Optional, trial billing cycle, for example "$1.99 for the first six months".
+    @available(iOS 15.0, *)
+    var trialBilling: PKRecurringPaymentSummaryItem? {
+        guard let dictionary = dict[ApplePayKeys.Recurring.trialBilling] as? [String: Any] else {
+            return nil
+        }
+        return .init(dictionary)
+    }
+
+        // Optional, localized billing agreement to be displayed to the user prior to payment authorization.
+    var billingAgreement: String? {
+        return dict[ApplePayKeys.Recurring.billingAgreement] as? String
+    }
+
+        // A URL that links to a page on your web site where the user can manage the payment method for this
+        // recurring payment, including deleting it.
+    var managementURL: URL? {
+        guard let url = dict[ApplePayKeys.Recurring.managementURL] as? String else { return nil }
+        return URL(string: url)
+    }
+
+    // Optional URL to receive lifecycle notifications for the merchant-specific payment token issued
+    // for this request, if applicable. If this property is not set, notifications will not be sent when
+    // lifecycle changes occur for the token, for example when the token is deleted.
+    var tokenNotificationURL: URL? {
+        guard let url = dict[ApplePayKeys.Recurring.tokenNotificationURL] as? String else { return nil }
+        return URL(string: url)
+    }
+
     public func buildConfiguration(payment: Payment) throws -> Adyen.ApplePayComponent.Configuration {
         let paymentRequest = try buildPaymentRequest(payment: payment)
         return try .init(paymentRequest: paymentRequest, allowOnboarding: allowOnboarding)
@@ -141,7 +184,19 @@ public struct ApplepayConfigurationParser {
         paymentRequest.shippingType = shippingType ?? .shipping
         paymentRequest.supportedCountries = supportedCountries
         paymentRequest.shippingMethods = shippingMethods
+        
+        guard #available(iOS 16.0, *), let paymentDescription, let regularBilling, let managementURL else {
+            return paymentRequest
+        }
 
+        let recurringRequest = PKRecurringPaymentRequest(paymentDescription: paymentDescription,
+                                                         regularBilling: regularBilling,
+                                                         managementURL: managementURL)
+        recurringRequest.tokenNotificationURL = tokenNotificationURL
+        recurringRequest.trialBilling = trialBilling
+        recurringRequest.billingAgreement = billingAgreement
+
+        paymentRequest.recurringPaymentRequest = recurringRequest
         return paymentRequest
     }
 }
