+++ ios/Configuration/Parameters.swift
@@ -71,6 +71,16 @@ internal enum ApplePayKeys: SubConfig {
     static var supportedCountries = "supportedCountries"
     static var shippingMethods = "shippingMethods"
 
+    enum Recurring {
+        static var paymentDescription = "description"
+        static var regularBilling = "regularBilling"
+        static var managementURL = "managementURL"
+
+        static var billingAgreement = "billingAgreement"
+        static var trialBilling = "trialBilling"
+        static var tokenNotificationURL = "tokenNotificationURL"
+    }
+
     enum Contact {
         static var phoneNumber = "phoneNumber"
         static var emailAddress = "emailAddress"
@@ -100,6 +110,36 @@ internal enum ApplePayKeys: SubConfig {
         static var dateComponentsRange = "dateComponentsRange"
         static var startDate = "startDate"
         static var endDate = "endDate"
+    }
+
+    enum RecurringPaymentSummaryItem {
+        static var label = "label"
+        static var amount = "amount"
+        static var type = "type"
+
+        static var startDate = "startDate"
+        static var endDate = "endDate"
+        static var intervalUnit = "intervalUnit"
+        static var intervalCount = "intervalCount"
+    }
+
+    enum CalendarUnit: String {
+        case minute, hour, day, month, year
+
+        var systemValue: NSCalendar.Unit {
+            switch self {
+            case .minute:
+                return .minute
+            case .hour:
+                return .hour
+            case .day:
+                return .day
+            case .month:
+                return .month
+            case .year:
+                return .year
+            }
+        }
     }
 }
